Create a solo ExternalBillingInputScreen to be used externally
